feat(ode): land collision_util / matrix / odemisc / timer + fast* C helpers
Wire six more ODE sources into both backends — collision_util,
matrix, odemisc, timer plus the four fastdot/fastldlt/fastlsolve/
fastltsolve C files matrix.cpp links against (dDot, dSolveL1,
dSolveL1T, dFactorLDLT).

Source fixes for -Werror clean build:
  - Widen `size_t maxl` to `int` in timer.cpp's per-event description
    width tracking so it matches the `%-*s` printf field-width
    parameter type (clang -Wformat).

lcp.cpp / stack.cpp still pull `dStopwatch` / `dStack` from gated
upstream headers; land with broader gated-ODE batch.

+++ src/physics/ode/timer.cpp
@@ -343,7 +343,7 @@ static void fprintDoubleWithPrefix (FILE *f, double a, const char *fmt)
 void dTimerReport (FILE *fout, int average)
 {
   int i;
-  size_t maxl;
+  int maxl;
   double ccunit = 1.0/dTimerTicksPerSecond();
   fprintf (fout,"\nTimer Report (");
   fprintDoubleWithPrefix (fout,ccunit,"%.2f ");
@@ -353,7 +353,7 @@ void dTimerReport (FILE *fout, int average)
   // get maximum description length
   maxl = 0;
   for (i=0; i<num; i++) {
-    size_t l = strlen (event[i].description);
+    int l = (int)strlen (event[i].description);
     if (l > maxl) maxl = l;
   }
